Update models and dbContext to use new default value attributes.

+++ Folly.Domain/Extensions/EntityEntryExtensions.cs
@@ -3,8 +3,8 @@ using Microsoft.EntityFrameworkCore.ChangeTracking;
 namespace Folly.Domain.Extensions;
 
 public static class EntityEntryExtensions {
-    public static long? GetPrimaryKey(this EntityEntry entry) {
-        if (long.TryParse(entry.Properties.FirstOrDefault(x => x.Metadata.IsPrimaryKey())?.CurrentValue?.ToString() ?? "", out var primaryKey)) {
+    public static int? GetPrimaryKey(this EntityEntry entry) {
+        if (int.TryParse(entry.Properties.FirstOrDefault(x => x.Metadata.IsPrimaryKey())?.CurrentValue?.ToString() ?? "", out var primaryKey)) {
             return primaryKey;
         }
         return null;
